requirements.txt
aws-cdk-lib==2.111.0
aws-cdk.aws-lambda-python-alpha==2.111.0a0
constructs==10.3.0
aws-lambda-powertools==2.26.0
boto3==1.33.6
